Refactor user service and types into folder
