+++ app/src/services/user/index.ts
@@ -1,0 +1,4 @@
+import { getUser, updateUser } from "./User";
+import { getConnections, removeConnection } from "./Connections";
+
+export { getUser, updateUser, getConnections, removeConnection };
